requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.7
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.5
pytz==2022.5
sqlparse==0.4.3
whitenoise==6.2.0
zipp==3.10.0
